changes in the discount logic

+++ utils/isValidNumber.js
@@ -1,4 +1,4 @@
-module.exports.isValidNumber=(number)=>{
-    number=parseInt(number);
-    return typeof number=== "number"&& number<=100 &&number >=0;
+module.exports.isValidNumber = (number) => {
+    number = parseFloat(number);
+    return typeof number === "number" && number <= 100 && number >= 0;
 }
